[outputs] Move device_add/rm to outputs and get rid of advertised flag
