+++ src/outputs/pulse.c
@@ -428,7 +428,6 @@ sinklist_cb(pa_context *ctx, const pa_sink_info *info, int eol, void *userdata)
   device->name = strdup(name);
   device->type = OUTPUT_TYPE_PULSE;
   device->type_name = outputs_name(device->type);
-  device->advertised = 1;
   device->extra_device_info = strdup(info->name);
 
   player_device_add(device);
